refactor: modularize drop part markdown
